+++ src/hydrogen/module/core.clj
@@ -58,7 +58,8 @@
         port (get-in options [:figwheel-main :port] 3449)
         host (get-in options [:figwheel-main :host] "0.0.0.0")
         watch-dirs (get-in options [:figwheel-main :watch-dirs])
-        reload-clj-files (get-in options [:figwheel-main :reload-clj-files])]
+        reload-clj-files (get-in options [:figwheel-main :reload-clj-files])
+        helpful-classpaths (get-in options [:figwheel-main :helpful-classpaths])]
     {:id "dev"
      :options (duct-server-figwheel-build-options config options)
      :config (cond-> {:mode :serve
@@ -69,7 +70,10 @@
                (assoc :watch-dirs watch-dirs)
 
                (some? reload-clj-files)
-               (assoc :reload-clj-files reload-clj-files))}))
+               (assoc :reload-clj-files reload-clj-files)
+
+               (some? helpful-classpaths)
+               (assoc :helpful-classpaths helpful-classpaths))}))
 
 (defn- core-config [config options]
   (let [environment (util/get-environment config options)]